Refactor plugin messaging
