+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/PacketListener.java
@@ -1,4 +1,4 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework;
+package com.craftmend.openaudiomc.api.velocitypluginmessageframework;
 
 /**
  * Classes implementing this interface MUST have an empty constructor (no parameters) in order for the packet api
