+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/PacketPlayer.java
@@ -1,4 +1,4 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework;
+package com.craftmend.openaudiomc.api.velocitypluginmessageframework;
 
 import lombok.Getter;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
