+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/StandardPacket.java
@@ -1,4 +1,4 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework;
+package com.craftmend.openaudiomc.api.velocitypluginmessageframework;
 
 import java.io.DataInputStream;
 import java.io.IOException;
