+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/bungeecord/PacketPlayerList.java
@@ -1,18 +1,18 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework.bungeecord;
+package com.craftmend.openaudiomc.libs.velocitypluginmessageframework.bungeecord;
 
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketWriter;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.RawPacket;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.PacketWriter;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.RawPacket;
 
 import java.io.IOException;
 
 /**
  * Created by iKeirNez on 01/01/14, ported to velocity and modified by fluse1367 on 11/2020.
  */
-public class PacketPlayerCount extends RawPacket {
+public class PacketPlayerList extends RawPacket {
 
     public String server;
 
-    public PacketPlayerCount(String server){
+    public PacketPlayerList(String server){
         super("BungeeCord");
         this.server = server;
     }
@@ -20,7 +20,7 @@ public class PacketPlayerCount extends RawPacket {
     @Override
     public PacketWriter write() throws IOException {
         PacketWriter packetWriter = new PacketWriter(this);
-        packetWriter.writeUTF("PlayerCount");
+        packetWriter.writeUTF("PlayerList");
         packetWriter.writeUTF(server);
         return packetWriter;
     }
