+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/bungeecord/PacketUUID.java
@@ -1,7 +1,7 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework.bungeecord;
+package com.craftmend.openaudiomc.libs.velocitypluginmessageframework.bungeecord;
 
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketWriter;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.RawPacket;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.PacketWriter;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.RawPacket;
 
 import java.io.IOException;
