+++ plugin/src/main/java/com/craftmend/openaudiomc/api/velocitypluginmessageframework/implementations/BungeeCordPacketManager.java
@@ -1,7 +1,7 @@
-package me.fluse1367.port.com.ikeirnez.pluginmessageframework.implementations;
+package com.craftmend.openaudiomc.libs.velocitypluginmessageframework.implementations;
 
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketManager;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketPlayer;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.PacketManager;
+import com.craftmend.openaudiomc.libs.velocitypluginmessageframework.PacketPlayer;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.connection.Server;
