+++ plugin/src/main/java/com/craftmend/openaudiomc/bungee/modules/node/NodeManager.java
@@ -2,8 +2,8 @@ package com.craftmend.openaudiomc.bungee.modules.node;
 
 import com.craftmend.openaudiomc.bungee.OpenAudioMcBungee;
 import com.craftmend.openaudiomc.bungee.modules.node.listeners.NodePacketListener;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketManager;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.implementations.BungeeCordPacketManager;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketManager;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.implementations.BungeeCordPacketManager;
 import lombok.Getter;
 
 public class NodeManager {
