+++ plugin/src/main/java/com/craftmend/openaudiomc/bungee/modules/player/listeners/PlayerConnectionListener.java
@@ -7,8 +7,8 @@ import com.craftmend.openaudiomc.generic.networking.packets.client.media.PacketC
 import com.craftmend.openaudiomc.generic.node.packets.ClientConnectedPacket;
 import com.craftmend.openaudiomc.generic.node.packets.ClientDisconnectedPacket;
 import com.craftmend.openaudiomc.generic.node.packets.ClientSyncHueStatePacket;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketPlayer;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.StandardPacket;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketPlayer;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.StandardPacket;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.event.PlayerDisconnectEvent;
 import net.md_5.bungee.api.event.PostLoginEvent;
