+++ plugin/src/main/java/com/craftmend/openaudiomc/generic/networking/client/objects/player/ClientConnection.java
@@ -28,7 +28,7 @@ import com.craftmend.openaudiomc.spigot.OpenAudioMcSpigot;
 import com.craftmend.openaudiomc.spigot.modules.proxy.enums.ClientMode;
 import com.craftmend.openaudiomc.velocity.OpenAudioMcVelocity;
 import com.craftmend.openaudiomc.velocity.generic.player.VelocityPlayerAdapter;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketPlayer;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketPlayer;
 import com.velocitypowered.api.proxy.Player;
 import lombok.Getter;
 import lombok.Setter;
