+++ plugin/src/main/java/com/craftmend/openaudiomc/generic/node/packets/ClientSyncHueStatePacket.java
@@ -1,8 +1,8 @@
 package com.craftmend.openaudiomc.generic.node.packets;
 
 import com.craftmend.openaudiomc.OpenAudioMc;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketWriter;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.StandardPacket;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketWriter;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.StandardPacket;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
