+++ plugin/src/main/java/com/craftmend/openaudiomc/velocity/modules/node/listeners/NodePacketListener.java
@@ -3,8 +3,8 @@ package com.craftmend.openaudiomc.velocity.modules.node.listeners;
 import com.craftmend.openaudiomc.OpenAudioMc;
 import com.craftmend.openaudiomc.generic.networking.client.objects.player.ClientConnection;
 import com.craftmend.openaudiomc.generic.node.packets.ForwardSocketPacket;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketHandler;
-import me.fluse1367.port.com.ikeirnez.pluginmessageframework.PacketListener;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketHandler;
+import com.craftmend.openaudiomc.api.velocitypluginmessageframework.PacketListener;
 
 import java.util.UUID;
